+++ ximpol/srcmodel/img.py
@@ -100,14 +100,23 @@ class xFITSImage:
         """
         return self.data[i][j]
 
-    def plot(self, show=True, zlabel='Counts/pixel',subplot=(1,1,1)):
+    def plot(self, show=True, zlabel='Counts/pixel', subplot=(1, 1, 1)):
         """Plot the image.
 
         This is using aplpy to render the image.
+
+        Warning
+        -------
+        We have to figure out the subplot issue, here. I put in a horrible
+        hack to recover the previous behavior when there's only one
+        subplot.
         """
         import aplpy
         with context_no_grids():
-            fig = aplpy.FITSFigure(self.hdu_list[0], figure=plt.figure(0,figsize=(10*subplot[1],10*subplot[0])), subplot = subplot)
+            if subplot == (1, 1, 1):
+                fig = aplpy.FITSFigure(self.hdu_list[0])
+            else:
+                fig = aplpy.FITSFigure(self.hdu_list[0], figure=plt.figure(0,figsize=(10*subplot[1], 10*subplot[0])), subplot=subplot)
         fig.add_grid()
         fig.show_colorscale(cmap = 'afmhot')
         fig.add_colorbar()